+++ cxx-qt-gen/test_outputs/handlers.cpp
@@ -1,4 +1,3 @@
-#include "cxx-qt-gen/include/my_object.cxx.h"
 #include "cxx-qt-gen/include/my_object.cxxqt.h"
 
 namespace cxx_qt::my_object {
