+++ cxx-qt-gen/test_outputs/handlers.rs
@@ -2,6 +2,9 @@
 mod my_object {
     unsafe extern "C++" {
         include!("cxx-qt-gen/include/my_object.cxxqt.h");
+        include!("cxx-qt-lib/include/convert.h");
+        include!("cxx-qt-lib/include/update_requester.h");
+        include ! (< QtCore / QObject >);
 
         #[cxx_name = "MyObject"]
         type MyObjectQt;
