+++ cxx-qt-gen/test_outputs/signals.h
@@ -1,15 +1,15 @@
 #pragma once
 
+#include <memory>
 #include <mutex>
 
-#include "cxx-qt-lib/include/qt_types.h"
+namespace cxx_qt::my_object {
+class MyObject;
+} // namespace cxx_qt::my_object
 
-#include <QtCore/QPoint>
-#include <QtCore/QVariant>
+#include "cxx-qt-gen/include/my_object.cxx.h"
 
 namespace cxx_qt::my_object {
-
-class MyObjectRust;
 
 class MyObject : public QObject
 {
