+++ cxx-qt-lib/build.rs
@@ -18,25 +18,34 @@ fn main() {
     // Copying the header is only needed for making the header available to a C++
     // build system, in which case CARGO_TARGET_DIR will be set by
     // the C++ build system.
+    println!("cargo:rerun-if-changed=include/convert.h");
     println!("cargo:rerun-if-changed=include/qt_types.h");
+    println!("cargo:rerun-if-changed=include/update_requester.h");
     println!("cargo:rerun-if-env-changed=CARGO_TARGET_DIR");
     if let Ok(target_dir) = env::var("CARGO_TARGET_DIR") {
         let manifest_dir = env::var("CARGO_MANIFEST_DIR").unwrap();
         std::fs::create_dir_all(&format!("{}/cxxbridge/cxx-qt-lib/include", target_dir)).unwrap();
-        // FIXME: Horrible hack around this sometimes failing because
-        // Windows doesn't allow multiple processes to access a file by default.
-        loop {
-            match std::fs::copy(
-                &format!("{}/include/qt_types.h", manifest_dir),
-                &format!("{}/cxxbridge/cxx-qt-lib/include/qt_types.h", target_dir),
-            ) {
-                Ok(_) => break,
-                #[cfg(windows)]
-                Err(e) if e.raw_os_error() == Some(32) => {
-                    std::thread::sleep(std::time::Duration::from_millis(100));
-                    continue;
+
+        for cpp_file in [
+            "include/convert.h",
+            "include/qt_types.h",
+            "include/update_requester.h",
+        ] {
+            // FIXME: Horrible hack around this sometimes failing because
+            // Windows doesn't allow multiple processes to access a file by default.
+            loop {
+                match std::fs::copy(
+                    &format!("{}/{}", manifest_dir, cpp_file),
+                    &format!("{}/cxxbridge/cxx-qt-lib/{}", target_dir, cpp_file),
+                ) {
+                    Ok(_) => break,
+                    #[cfg(windows)]
+                    Err(e) if e.raw_os_error() == Some(32) => {
+                        std::thread::sleep(std::time::Duration::from_millis(100));
+                        continue;
+                    }
+                    Err(e) => panic!("Error copying {}: {:?}", e, cpp_file),
                 }
-                Err(e) => panic!("Error copying qt_types.h: {:?}", e),
             }
         }
     }
@@ -68,7 +77,7 @@ fn main() {
     }
 
     let mut builder = cxx_build::bridges(&bridge_files);
-    for cpp_file in ["src/qt_types.cpp"] {
+    for cpp_file in ["src/qt_types.cpp", "src/update_requester.cpp"] {
         builder.file(cpp_file);
         println!("cargo:rerun-if-changed={}", cpp_file);
     }
