+++ cxx-qt-lib/include/qt_types.h
@@ -1,6 +1,7 @@
-// SPDX-FileCopyrightText: 2022 Klarälvdalens Datakonsult AB, a KDAB Group
-// company <info@kdab.com> SPDX-FileContributor: Andrew Hayzen
-// <andrew.hayzen@kdab.com>
+// clang-format off
+// SPDX-FileCopyrightText: 2022 Klarälvdalens Datakonsult AB, a KDAB Group company <info@kdab.com>
+// clang-format on
+// SPDX-FileContributor: Andrew Hayzen <andrew.hayzen@kdab.com>
 //
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
@@ -12,7 +13,6 @@
 #include <QDateTime>
 #include <QPoint>
 #include <QPointF>
-#include <QPointer>
 #include <QRect>
 #include <QRectF>
 #include <QSize>
@@ -26,59 +26,6 @@
 
 namespace rust {
 namespace cxxqtlib1 {
-
-template<typename R, typename T>
-struct cxx_qt_convert
-{
-  static_assert(std::is_convertible_v<T, R>, R"(
-
-CXXQt: No viable conversion between types found.
-Consider defining your own type conversion.
-See: https://kdab.github.io/cxx-qt/book/concepts/type-conversions.html
-)");
-  R operator()(T val) { return val; }
-};
-
-template<typename R, typename T>
-struct cxx_qt_convert<R&, T>
-{
-  R& operator()(T& val) { return val; }
-};
-
-template<typename R, typename T>
-struct cxx_qt_convert<const R&, T>
-{
-  const R& operator()(const T& val) { return val; }
-};
-
-template<typename R, typename T>
-struct cxx_qt_convert<R, std::unique_ptr<T>>
-{
-  R operator()(std::unique_ptr<T> ptr) { return std::move(*ptr); }
-};
-
-template<typename R, typename T>
-struct cxx_qt_convert<R&, std::unique_ptr<T>>
-{
-  R& operator()(std::unique_ptr<T>& ptr) { return *ptr; }
-};
-
-template<typename R, typename T>
-struct cxx_qt_convert<const R&, std::unique_ptr<T>>
-{
-  const R& operator()(const std::unique_ptr<T>& ptr) { return *ptr; }
-};
-
-class UpdateRequester
-{
-public:
-  UpdateRequester(QPointer<QObject> obj, const char* method);
-  bool requestUpdate() const;
-
-private:
-  const char* m_method;
-  QPointer<QObject> m_obj;
-};
 
 namespace types {
