+++ cxx-qt-lib/src/qt_types.cpp
@@ -1,6 +1,7 @@
-// SPDX-FileCopyrightText: 2022 Klarälvdalens Datakonsult AB, a KDAB Group
-// company <info@kdab.com> SPDX-FileContributor: Andrew Hayzen
-// <andrew.hayzen@kdab.com>
+// clang-format off
+// SPDX-FileCopyrightText: 2022 Klarälvdalens Datakonsult AB, a KDAB Group company <info@kdab.com>
+// clang-format on
+// SPDX-FileContributor: Andrew Hayzen <andrew.hayzen@kdab.com>
 //
 // SPDX-License-Identifier: MIT OR Apache-2.0
 
@@ -10,22 +11,6 @@
 
 namespace rust {
 namespace cxxqtlib1 {
-
-UpdateRequester::UpdateRequester(QPointer<QObject> obj, const char* method)
-  : m_method(method)
-  , m_obj(obj)
-{
-}
-
-bool
-UpdateRequester::requestUpdate() const
-{
-  if (m_obj == nullptr) {
-    return false;
-  }
-
-  return QMetaObject::invokeMethod(m_obj, m_method, Qt::QueuedConnection);
-}
 
 std::unique_ptr<QColor>
 qcolorInit()
@@ -366,5 +351,5 @@ CXX_QT_VARIANT_TRIVIAL_VALUE(quint8, U8)
 CXX_QT_VARIANT_TRIVIAL_VALUE(quint16, U16)
 CXX_QT_VARIANT_TRIVIAL_VALUE(quint32, U32)
 
-}
-}
+} // namespace cxxqtlib1
+} // namespace rust
