+++ tests/basic_cxx_only/include/main.h
@@ -7,5 +7,7 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 #pragma once
 
+#include "cxx-qt-gen/include/ffi.cxx.h"
+
 int
 get_cpp_number();
