+++ tests/basic_cxx_only/src/main.cpp
@@ -8,7 +8,6 @@
 #define DOCTEST_CONFIG_IMPLEMENT_WITH_MAIN
 #include "doctest.h"
 
-#include "cxx-qt-gen/include/ffi.cxx.h"
 #include "main.h"
 
 int hidden_num = 100;
